Fixes for two exceptions in the view
a) Strict standards: Declaration of Slimx\View::render() should be compatible with Slim\View::render($template, $data = NULL)

Fixed the signature and also added handling for the passed-in $data array.

b) Fatal error: Call to a member function set() on a non-object

The parent view constructor was never invoked and thus `$this->data = new \Slim\Helper\Set();` was never called.

I tested this with Slim 2.5. Not sure if this is the reason why it didn't work.